FIX collate logics.
The old default collate_fn adds batch dimension at the end,
while it is hard to split data further in each process, especially
when data contains string in which case the data is of hybrid type
and there is no permute() available.

The old code also has wrong collate logics.